refactor(mir): split ownership and scope lowering methods
I move the ownership and scope-exit Builder methods into dedicated lower modules while preserving their bodies and order. I also keep the structural source inventory aligned with the new module boundary.
